Fixing casting

+++ app/src/main/java/dk/ucn/dmaa0214/group1/meteringapp/TiltAngleActivityFragment.java
@@ -49,7 +49,7 @@ public class TiltAngleActivityFragment extends Fragment {
             public void onSensorChanged(SensorEvent event) {
                 float[] values = fixValues(event);
 
-                int i = (int) (double) (int) ((182D * Math.atan(values[0] / values[1]) / 3.1415926535897931D));
+                int i = (int) ((182D * Math.atan(values[0] / values[1]) / 3.1415926535897931D));
 
                 int l = i;
                 int j = l;
